readme cleanup and removed node specific funtions from deployment.ts and proving.ts
