+++ backend/contracts/aztec/L2AztecBridgeAdapter/src/artifacts/L2AztecBridgeAdapter.ts
@@ -33,9 +33,10 @@ import {
   type Wallet,
   type U128Like,
   type WrappedFieldLike,
+  //@ts-ignore
 } from '@aztec/aztec.js';
 //@ts-ignore
-import L2AztecBridgeAdapterContractArtifactJson from '../../target/L2AztecBridgeAdapter-L2AztecBridgeAdapter.json' with {type:"json"};
+import L2AztecBridgeAdapterContractArtifactJson from '../../target/L2AztecBridgeAdapter-L2AztecBridgeAdapter.json' with { type: 'json' };
 export const L2AztecBridgeAdapterContractArtifact = loadContractArtifact(L2AztecBridgeAdapterContractArtifactJson as NoirCompiledContract);
