+++ backend/contracts/aztec/WarpToadCore/src/artifacts/WarpToadCore.ts
@@ -33,9 +33,10 @@ import {
   type Wallet,
   type U128Like,
   type WrappedFieldLike,
+  //@ts-ignore
 } from '@aztec/aztec.js';
 //@ts-ignore
-import WarpToadCoreContractArtifactJson from '../../target/WarpToadCore-WarpToadCore.json'  with {type:"json"};
+import WarpToadCoreContractArtifactJson from '../../target/WarpToadCore-WarpToadCore.json' with { type: 'json' };
 export const WarpToadCoreContractArtifact = loadContractArtifact(WarpToadCoreContractArtifactJson as NoirCompiledContract);
