+++ backend/package.json
@@ -1,4 +1,62 @@
 {
+  "name": "@warp-toad/backend",
+  "private": true,
+  "version": "1.0.18",
+  "repository": "https://github.com/warptoad/warp-toad/tree/backend-npm-package/backend",
+  "author": "Jim Jim Valkema",
+  "license": "MIT",
+  "scripts": {
+    "build": "yarn hardhat clean; yarn hardhat compile; tsc --build tsconfigOnBuild.json; cp -r ignition/deployments/ dist/ignition/; cp -r scripts/deploy/aztec/aztecDeployments/ dist/scripts/deploy/aztec/"
+  },
+  "devDependencies": {
+    "@nomicfoundation/hardhat-chai-matchers": "^2.0.8",
+    "@nomicfoundation/hardhat-ethers": "^3.0.8",
+    "@nomicfoundation/hardhat-ignition": "^0.15.10",
+    "@nomicfoundation/hardhat-ignition-ethers": "^0.15.10",
+    "@nomicfoundation/hardhat-network-helpers": "^1.0.12",
+    "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+    "@nomicfoundation/hardhat-verify": "^2.0.14",
+    "@nomicfoundation/ignition-core": "^0.15.10",
+    "@typechain/ethers-v6": "^0.5.1",
+    "@typechain/hardhat": "^9.1.0",
+    "@types/chai": "4.3.20",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^24.0.10",
+    "chai": "4.5.0",
+    "ethers": "^6.13.5",
+    "hardhat": "^2.26.0",
+    "hardhat-gas-reporter": "1.0.10",
+    "jest": "^29.7.0",
+    "patch-package": "^8.0.0",
+    "readline": "^1.3.0",
+    "solidity-coverage": "^0.8.14",
+    "ts-node": "^10.9.2",
+    "typechain": "^8.3.2",
+    "typescript": "^5.8.2"
+  },
+  "dependencies": {
+    "@aztec/accounts": "1.1.3",
+    "@aztec/aztec.js": "1.1.3",
+    "@aztec/bb.js": "0.72.1",
+    "@aztec/ethereum": "1.1.3",
+    "@aztec/foundation": "1.1.3",
+    "@aztec/noir-contracts.js": "1.1.3",
+    "@noir-lang/noir_js": "1.0.0-beta.5",
+    "@noir-lang/noir_wasm": "1.0.0-beta.5",
+    "@openzeppelin/contracts": "^5.2.0",
+    "@scroll-tech/contracts": "^2.0.1",
+    "@types/argparse": "^2.0.17",
+    "@zk-kit/imt": "^2.0.0-beta.8",
+    "@zk-kit/lazy-imt.sol": "^2.0.0-beta.12",
+    "@zkpassport/poseidon2": "^0.6.1",
+    "argparse": "^2.0.1",
+    "bun": "^1.2.12",
+    "fixed-merkle-tree": "^0.7.3",
+    "minimatch": "^10.0.3",
+    "poseidon-lite": "^0.3.0",
+    "ts-node": "^10.9.2",
+    "warp-toad-old-backend": "^1.0.18"
+  },
   "main": "./dist/index.js",
   "esbuildOptions": {
     "target": "esnext"
@@ -99,72 +157,14 @@
     },
     "./deployment": {
       "import": {
-        "types": "./dist/scripts/dev_op/utils.d.ts",
-        "default": "./dist/scripts/dev_op/utils.js"
+        "types": "./dist/scripts/dev_op/deployment.d.ts",
+        "default": "./dist/scripts/dev_op/deployment.js"
       },
       "require": {
-        "types": "./dist/scripts/dev_op/utils.d.ts",
-        "default": "./dist/scripts/dev_op/utils.js"
+        "types": "./dist/scripts/dev_op/deployment.d.ts",
+        "default": "./dist/scripts/dev_op/deployment.js"
       }
     }
-  },
-  "name": "warp-toad-old-backend",
-  "private": false,
-  "version": "1.0.18",
-  "repository": "https://github.com/warptoad/warp-toad/tree/backend-npm-package/backend",
-  "author": "Jim Jim Valkema",
-  "license": "MIT",
-  "scripts": {
-    "build": "yarn tsc --build tsconfigOnBuild.json; cp -r ignition/deployments/ dist/ignition/; cp -r scripts/deploy/aztec/aztecDeployments/ dist/scripts/deploy/aztec/"
-  },
-  "devDependencies": {
-    "@nomicfoundation/hardhat-chai-matchers": "^2.0.8",
-    "@nomicfoundation/hardhat-ethers": "^3.0.8",
-    "@nomicfoundation/hardhat-ignition": "^0.15.10",
-    "@nomicfoundation/hardhat-ignition-ethers": "^0.15.10",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.12",
-    "@nomicfoundation/hardhat-toolbox": "^5.0.0",
-    "@nomicfoundation/hardhat-verify": "^2.0.14",
-    "@nomicfoundation/ignition-core": "^0.15.10",
-    "@typechain/ethers-v6": "^0.5.1",
-    "@typechain/hardhat": "^9.1.0",
-    "@types/chai": "4.3.20",
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^24.0.10",
-    "chai": "4.5.0",
-    "ethers": "^6.13.5",
-    "hardhat": "2.24.1",
-    "hardhat-gas-reporter": "1.0.10",
-    "jest": "^29.7.0",
-    "patch-package": "^8.0.0",
-    "readline": "^1.3.0",
-    "solidity-coverage": "^0.8.14",
-    "ts-node": "^10.9.2",
-    "typechain": "^8.3.2",
-    "typescript": "^5.8.2"
-  },
-  "dependencies": {
-    "@aztec/accounts": "1.1.3",
-    "@aztec/aztec.js": "1.1.3",
-    "@aztec/bb.js": "0.72.1",
-    "@aztec/ethereum": "1.1.3",
-    "@aztec/foundation": "1.1.3",
-    "@aztec/noir-contracts.js": "1.1.3",
-    "@noir-lang/noir_js": "1.0.0-beta.5",
-    "@noir-lang/noir_wasm": "1.0.0-beta.5",
-    "@openzeppelin/contracts": "^5.2.0",
-    "@scroll-tech/contracts": "^2.0.1",
-    "@types/argparse": "^2.0.17",
-    "@zk-kit/imt": "^2.0.0-beta.8",
-    "@zk-kit/lazy-imt.sol": "^2.0.0-beta.12",
-    "@zkpassport/poseidon2": "^0.6.1",
-    "argparse": "^2.0.1",
-    "bun": "^1.2.12",
-    "fixed-merkle-tree": "^0.7.3",
-    "minimatch": "^10.0.3",
-    "poseidon-lite": "^0.3.0",
-    "ts-node": "^10.9.2",
-    "warp-toad-old-backend": "1.0.18"
   },
   "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
