+++ backend/scripts/deploy/L1/deployL1.ts
@@ -12,7 +12,8 @@ import { ERC20__factory, USDcoin__factory } from "../../../typechain-types";
 import er20Abi from "../../dev_op/erc20ABI.json"  with { type: 'json' }
 import { L1_SCROLL_MESSENGER_MAINNET, L1_SCROLL_MESSENGER_SEPOLIA } from "../../lib/constants";
 import fs from "fs/promises";
-import { checkFileExists, getContractAddressesEvm, getEvmDeployedAddressesFilePath, getEvmDeployedAddressesFolderPath, promptBool } from "../../dev_op/utils";
+import { getContractAddressesEvm, getEvmDeployedAddressesFilePath, getEvmDeployedAddressesFolderPath } from "../../dev_op/deployment";
+import { checkFileExists, promptBool } from "scripts/dev_op/utils";
 const sleep = (ms: number) => new Promise(r => setTimeout(r, ms));
