+++ backend/scripts/deploy/L1/initializeL1.ts
@@ -16,7 +16,7 @@ import { createPXEClient, waitForPXE } from "@aztec/aztec.js";
 import { getInitialTestAccountsWallets } from "@aztec/accounts/testing";
 
 import fs from "fs/promises";
-import { getContractAddressesAztec, getContractAddressesEvm } from "../../dev_op/utils";
+import { getContractAddressesAztec, getContractAddressesEvm } from "../../dev_op/deployment";
 
 
 function getArgs() {
