+++ backend/scripts/deploy/aztec/deployAztec.ts
@@ -19,7 +19,7 @@ import er20Abi from "../../dev_op/erc20ABI.json"  with { type: 'json' }
 import { deployL2AztecBridgeAdapter } from "./L2AztecBridgeAdapter";
 
 import hre, { network } from "hardhat"
-import { checkFileExists, getAztecDeployedAddressesFilePath, getAztecDeployedAddressesFolderPath, getContractAddressesEvm, promptBool } from "../../dev_op/utils";
+import { getAztecDeployedAddressesFilePath, getAztecDeployedAddressesFolderPath, getContractAddressesEvm } from "../../dev_op/deployment";
 //@ts-ignore
 import { getSchnorrAccount } from "@aztec/accounts/schnorr";
 //@ts-ignore
@@ -31,9 +31,9 @@ import { SingleKeyAccountContract } from "@aztec/accounts/single_key";
 import { createAztecNodeClient } from "@aztec/stdlib/interfaces/client";
 //@ts-ignore
 import { computePartialAddress } from "@aztec/stdlib/contract";
-import { getAztecTestWallet } from "../../dev_op/utils";
+import { getAztecTestWallet } from "../../dev_op/deployment";
+import { checkFileExists, promptBool } from "scripts/dev_op/utils";
 //import { ObsidionDeployerFPCContractArtifact } from "../dev_op/getObsidionWallet/ObsidionDeployerFPC"
-
