+++ backend/scripts/deploy/scroll/initializeL2Scroll.ts
@@ -16,7 +16,7 @@ import { createPXEClient, waitForPXE } from "@aztec/aztec.js";
 import { getInitialTestAccountsWallets } from "@aztec/accounts/testing";
 
 import fs from "fs/promises";
-import { getContractAddressesAztec, getContractAddressesEvm } from "../../dev_op/utils";
+import { getContractAddressesAztec, getContractAddressesEvm } from "../../dev_op/deployment";
 import { vars } from "hardhat/config.js";
 const SEPOLIA_URL = vars.get("SEPOLIA_URL")
